requirements.txt
fastapi
uvicorn
jinja2
requests
mangum
psycopg2-binary